feat: create dtos for supabase objects
