--- chessevolved_model/src/main/kotlin/io/github/chessevolved/components/GameState.kt
@@ -1,8 +1,0 @@
-package io.github.chessevolved.components
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class GameState(
-    val pieces: List<SerializablePiece>,
-    val boardSquares: List<SerializableBoardSquare>,
-)
