--- chessevolved_model/src/main/kotlin/io/github/chessevolved/components/SerializableBoardSquare.kt
@@ -1,8 +1,0 @@
-package io.github.chessevolved.components
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SerializableBoardSquare(
-    val position: Position,
-    val weather: WeatherEvent,
-)
