--- chessevolved_model/src/main/kotlin/io/github/chessevolved/components/SerializablePiece.kt
@@ -1,9 +1,0 @@
-package io.github.chessevolved.components
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SerializablePiece(
-    val position: Position,
-    val type: PieceType,
-    val color: PlayerColor,
-)
