--- chessevolved_shared/src/main/kotlin/io/github/chessevolved_shared/SettingsDTO.kt
@@ -1,6 +1,0 @@
-package io.github.chessevolved.shared
-
-data class SettingsDTO(
-    val fogOfWar: Boolean,
-    val boardSize: Int,
-)
