+++ chessevolved_shared/src/main/kotlin/io/github/chessevolved_shared/dtos/BoardSquareDto.kt
@@ -1,0 +1,11 @@
+package io.github.chessevolved.dtos
+
+import io.github.chessevolved.components.Position
+import io.github.chessevolved.components.WeatherEvent
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BoardSquareDto(
+    val position: Position,
+    val weather: WeatherEvent,
+)
