+++ chessevolved_shared/src/main/kotlin/io/github/chessevolved_shared/dtos/LobbyDto.kt
@@ -1,0 +1,13 @@
+package io.github.chessevolved.dtos
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LobbyDto(
+    val id: Int,
+    val created_at: String,
+    val lobby_code: String,
+    val second_player: Boolean,
+    val game_started: Boolean,
+    val settings: List<SettingsDto>,
+)
