+++ chessevolved_shared/src/main/kotlin/io/github/chessevolved_shared/dtos/PieceDto.kt
@@ -1,0 +1,13 @@
+package io.github.chessevolved.dtos
+
+import io.github.chessevolved.components.PieceType
+import io.github.chessevolved.components.PlayerColor
+import io.github.chessevolved.components.Position
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PieceDto(
+    val position: Position,
+    val type: PieceType,
+    val color: PlayerColor,
+)
