+++ chessevolved_shared/src/main/kotlin/io/github/chessevolved_shared/dtos/SettingsDto.kt
@@ -1,0 +1,6 @@
+package io.github.chessevolved.dtos
+
+data class SettingsDto(
+    val fogOfWar: Boolean,
+    val boardSize: Int,
+)
